git-svn-id: svn://192.168.1.10/public/CoffeeMud@10519 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ guides/database/coffeemuddbmysql.sql
@@ -148,7 +148,7 @@ CREATE TABLE CMCHIT (
 	CMITID varchar (50) NULL ,
 	CMITTX longtext NULL ,
 	CMITLO varchar (100) NULL ,
-	CMITWO int NULL ,
+	CMITWO bigint NULL ,
 	CMITUR int NULL ,
 	CMITLV int NULL ,
 	CMITAB int NULL ,
